fix streaming binary stl that starts with 'solid'
This used to enter ascii mode and fail by not
emitting any verts.

Now there are two ways of catching this:

1. buffer data until the first 'facet' is reached.
   At that point, the file is "known" to be ascii.

2. the stream ends and a 'facet' still hasn't been
   found.

In either case we buffer data until it is known.

For ascii files the overhead should be negligable,
but in the binary case with no coincidental \n the
end result is the entire file is buffered before
processing.